Changes due to changes in JSBML.
git-svn-id: svn://rarepos.cs.uni-tuebingen.de/SysBio/trunk@85 e728d8a2-de87-4acf-9dc5-8dc47f4ff2f0

+++ src/de/zbit/kegg/gui/HTMLFormula.java
@@ -93,7 +93,7 @@ public class HTMLFormula extends StringTools implements ASTNodeCompiler {
 			times = HTMLFormula.sum(StringTools.toString(u.getOffset()),
 					times);
 		}
-		return HTMLFormula.pow(times, Integer.valueOf(u.getExponent())).toString();
+		return HTMLFormula.pow(times, Double.valueOf(u.getExponent())).toString();
 	}
